MMath.LargestDoubleSum fix

+++ src/Runtime/Core/Maths/SpecialFunctions.cs
@@ -4736,11 +4736,12 @@ else if (m < 20.0 - 60.0/11.0 * s) {
             double upperBound;
             if (Math.Abs(sum) > Math.Abs(b))
             {
-                upperBound = b + NextDouble(sum);
+                // Cast to double to prevent upperBound from landing between double.MaxValue and infinity on a 32-bit platform.
+                upperBound = (double)(b + NextDouble(sum));
             }
             else
             {
-                upperBound = NextDouble(b) + sum;
+                upperBound = (double)(NextDouble(b) + sum);
             }
             int iterCount = 0;
             while (true)
